nvif: split out fifo interface definitions
